planner: move exhaustPhysicalPlans4LogicalJoin into physical op dir
