+++ pkg/planner/cascades/old/implementation_rules.go
@@ -558,7 +558,7 @@ func (*ImplApply) Match(expr *memo.GroupExpr, prop *property.PhysicalProperty) (
 // OnImplement implements ImplementationRule OnImplement interface
 func (*ImplApply) OnImplement(expr *memo.GroupExpr, reqProp *property.PhysicalProperty) ([]memo.Implementation, error) {
 	la := expr.ExprNode.(*logicalop.LogicalApply)
-	join := plannercore.GetHashJoin(nil, la, reqProp)
+	join := physicalop.GetHashJoin(nil, la, reqProp)
 	physicalApply := physicalop.PhysicalApply{
 		PhysicalHashJoin: *join,
 		OuterSchema:      la.CorCols,
