+++ pkg/planner/core/operator/physicalop/physical_hash_join.go
@@ -16,9 +16,11 @@ package physicalop
 
 import (
 	"fmt"
+	"math"
 	"strings"
 
 	"github.com/pingcap/errors"
+	"github.com/pingcap/failpoint"
 	"github.com/pingcap/tidb/pkg/executor/join/joinversion"
 	"github.com/pingcap/tidb/pkg/expression"
 	"github.com/pingcap/tidb/pkg/kv"
@@ -28,6 +30,7 @@ import (
 	"github.com/pingcap/tidb/pkg/planner/util/costusage"
 	"github.com/pingcap/tidb/pkg/planner/util/optimizetrace"
 	"github.com/pingcap/tidb/pkg/planner/util/utilfuncp"
+	h "github.com/pingcap/tidb/pkg/util/hint"
 	"github.com/pingcap/tidb/pkg/util/plancodec"
 	"github.com/pingcap/tidb/pkg/util/size"
 	"github.com/pingcap/tipb/go-tipb"
@@ -624,4 +627,324 @@ func (p *PhysicalHashJoin) ToPB(ctx *base.BuildPBContext, storeType kv.StoreType
 		FineGrainedShuffleStreamCount: p.TiFlashFineGrainedShuffleStreamCount,
 		FineGrainedShuffleBatchSize:   ctx.TiFlashFineGrainedShuffleBatchSize,
 	}, nil
+}
+
+func tryToGetMppHashJoin(super base.LogicalPlan, prop *property.PhysicalProperty, useBCJ bool) []base.PhysicalPlan {
+	ge, p := base.GetGEAndLogical[*logicalop.LogicalJoin](super)
+	if !prop.IsSortItemEmpty() {
+		return nil
+	}
+	if prop.TaskTp != property.RootTaskType && prop.TaskTp != property.MppTaskType {
+		return nil
+	}
+
+	if !expression.IsPushDownEnabled(p.JoinType.String(), kv.TiFlash) {
+		p.SCtx().GetSessionVars().RaiseWarningWhenMPPEnforced("MPP mode may be blocked because join type `" + p.JoinType.String() + "` is blocked by blacklist, check `table mysql.expr_pushdown_blacklist;` for more information.")
+		return nil
+	}
+
+	if p.JoinType != base.InnerJoin && p.JoinType != base.LeftOuterJoin && p.JoinType != base.RightOuterJoin && p.JoinType != base.SemiJoin && p.JoinType != base.AntiSemiJoin && p.JoinType != base.LeftOuterSemiJoin && p.JoinType != base.AntiLeftOuterSemiJoin {
+		p.SCtx().GetSessionVars().RaiseWarningWhenMPPEnforced("MPP mode may be blocked because join type `" + p.JoinType.String() + "` is not supported now.")
+		return nil
+	}
+
+	if len(p.EqualConditions) == 0 {
+		if !useBCJ {
+			p.SCtx().GetSessionVars().RaiseWarningWhenMPPEnforced("MPP mode may be blocked because `Cartesian Product` is only supported by broadcast join, check value and documents of variables `tidb_broadcast_join_threshold_size` and `tidb_broadcast_join_threshold_count`.")
+			return nil
+		}
+		if p.SCtx().GetSessionVars().AllowCartesianBCJ == 0 {
+			p.SCtx().GetSessionVars().RaiseWarningWhenMPPEnforced("MPP mode may be blocked because `Cartesian Product` is only supported by broadcast join, check value and documents of variable `tidb_opt_broadcast_cartesian_join`.")
+			return nil
+		}
+	}
+	if len(p.LeftConditions) != 0 && p.JoinType != base.LeftOuterJoin {
+		p.SCtx().GetSessionVars().RaiseWarningWhenMPPEnforced("MPP mode may be blocked because there is a join that is not `left join` but has left conditions, which is not supported by mpp now, see github.com/pingcap/tidb/issues/26090 for more information.")
+		return nil
+	}
+	if len(p.RightConditions) != 0 && p.JoinType != base.RightOuterJoin {
+		p.SCtx().GetSessionVars().RaiseWarningWhenMPPEnforced("MPP mode may be blocked because there is a join that is not `right join` but has right conditions, which is not supported by mpp now.")
+		return nil
+	}
+
+	if prop.MPPPartitionTp == property.BroadcastType {
+		return nil
+	}
+	if !canExprsInJoinPushdown(p, kv.TiFlash) {
+		return nil
+	}
+	lkeys, rkeys, _, _ := p.GetJoinKeys()
+	lNAkeys, rNAKeys := p.GetNAJoinKeys()
+	// check match property
+	baseJoin := BasePhysicalJoin{
+		JoinType:        p.JoinType,
+		LeftConditions:  p.LeftConditions,
+		RightConditions: p.RightConditions,
+		OtherConditions: p.OtherConditions,
+		DefaultValues:   p.DefaultValues,
+		LeftJoinKeys:    lkeys,
+		RightJoinKeys:   rkeys,
+		LeftNAJoinKeys:  lNAkeys,
+		RightNAJoinKeys: rNAKeys,
+	}
+	// It indicates which side is the build side.
+	forceLeftToBuild := ((p.PreferJoinType & h.PreferLeftAsHJBuild) > 0) || ((p.PreferJoinType & h.PreferRightAsHJProbe) > 0)
+	forceRightToBuild := ((p.PreferJoinType & h.PreferRightAsHJBuild) > 0) || ((p.PreferJoinType & h.PreferLeftAsHJProbe) > 0)
+	if forceLeftToBuild && forceRightToBuild {
+		p.SCtx().GetSessionVars().StmtCtx.SetHintWarning(
+			"Some HASH_JOIN_BUILD and HASH_JOIN_PROBE hints are conflicts, please check the hints")
+		forceLeftToBuild = false
+		forceRightToBuild = false
+	}
+	preferredBuildIndex := 0
+	fixedBuildSide := false // Used to indicate whether the build side for the MPP join is fixed or not.
+	var stats0, stats1 *property.StatsInfo
+	if ge != nil {
+		stats0, stats1, _, _ = ge.GetJoinChildStatsAndSchema()
+	} else {
+		stats0, stats1, _, _ = p.GetJoinChildStatsAndSchema()
+	}
+	if p.JoinType == base.InnerJoin {
+		if stats0.Count() > stats1.Count() {
+			preferredBuildIndex = 1
+		}
+	} else if p.JoinType.IsSemiJoin() {
+		if !useBCJ && !p.IsNAAJ() && len(p.EqualConditions) > 0 && (p.JoinType == base.SemiJoin || p.JoinType == base.AntiSemiJoin) {
+			// TiFlash only supports Non-null_aware non-cross semi/anti_semi join to use both sides as build side
+			preferredBuildIndex = 1
+			// MPPOuterJoinFixedBuildSide default value is false
+			// use MPPOuterJoinFixedBuildSide here as a way to disable using left table as build side
+			if !p.SCtx().GetSessionVars().MPPOuterJoinFixedBuildSide && stats1.Count() > stats0.Count() {
+				preferredBuildIndex = 0
+			}
+		} else {
+			preferredBuildIndex = 1
+			fixedBuildSide = true
+		}
+	}
+	if p.JoinType == base.LeftOuterJoin || p.JoinType == base.RightOuterJoin {
+		// TiFlash does not require that the build side must be the inner table for outer join.
+		// so we can choose the build side based on the row count, except that:
+		// 1. it is a broadcast join(for broadcast join, it makes sense to use the broadcast side as the build side)
+		// 2. or session variable MPPOuterJoinFixedBuildSide is set to true
+		// 3. or nullAware/cross joins
+		if useBCJ || p.IsNAAJ() || len(p.EqualConditions) == 0 || p.SCtx().GetSessionVars().MPPOuterJoinFixedBuildSide {
+			if !p.SCtx().GetSessionVars().MPPOuterJoinFixedBuildSide {
+				// The hint has higher priority than variable.
+				fixedBuildSide = true
+			}
+			if p.JoinType == base.LeftOuterJoin {
+				preferredBuildIndex = 1
+			}
+		} else if stats0.Count() > stats1.Count() {
+			preferredBuildIndex = 1
+		}
+	}
+
+	if forceLeftToBuild || forceRightToBuild {
+		match := (forceLeftToBuild && preferredBuildIndex == 0) || (forceRightToBuild && preferredBuildIndex == 1)
+		if !match {
+			if fixedBuildSide {
+				// A warning will be generated if the build side is fixed, but we attempt to change it using the hint.
+				p.SCtx().GetSessionVars().StmtCtx.SetHintWarning(
+					"Some HASH_JOIN_BUILD and HASH_JOIN_PROBE hints cannot be utilized for MPP joins, please check the hints")
+			} else {
+				// The HASH_JOIN_BUILD OR HASH_JOIN_PROBE hints can take effective.
+				preferredBuildIndex = 1 - preferredBuildIndex
+			}
+		}
+	}
+
+	// set preferredBuildIndex for test
+	failpoint.Inject("mockPreferredBuildIndex", func(val failpoint.Value) {
+		if !p.SCtx().GetSessionVars().InRestrictedSQL {
+			preferredBuildIndex = val.(int)
+		}
+	})
+
+	baseJoin.InnerChildIdx = preferredBuildIndex
+	childrenProps := make([]*property.PhysicalProperty, 2)
+	if useBCJ {
+		childrenProps[preferredBuildIndex] = &property.PhysicalProperty{TaskTp: property.MppTaskType, ExpectedCnt: math.MaxFloat64, MPPPartitionTp: property.BroadcastType, CanAddEnforcer: true, CTEProducerStatus: prop.CTEProducerStatus}
+		expCnt := math.MaxFloat64
+		if prop.ExpectedCnt < p.StatsInfo().RowCount {
+			expCntScale := prop.ExpectedCnt / p.StatsInfo().RowCount
+			var targetStats *property.StatsInfo
+			if 1-preferredBuildIndex == 0 {
+				targetStats = stats0
+			} else {
+				targetStats = stats1
+			}
+			expCnt = targetStats.RowCount * expCntScale
+		}
+		if prop.MPPPartitionTp == property.HashType {
+			lPartitionKeys, rPartitionKeys := p.GetPotentialPartitionKeys()
+			hashKeys := rPartitionKeys
+			if preferredBuildIndex == 1 {
+				hashKeys = lPartitionKeys
+			}
+			matches := prop.IsSubsetOf(hashKeys)
+			if len(matches) == 0 {
+				return nil
+			}
+			childrenProps[1-preferredBuildIndex] = &property.PhysicalProperty{TaskTp: property.MppTaskType, ExpectedCnt: expCnt, MPPPartitionTp: property.HashType, MPPPartitionCols: prop.MPPPartitionCols, CTEProducerStatus: prop.CTEProducerStatus}
+		} else {
+			childrenProps[1-preferredBuildIndex] = &property.PhysicalProperty{TaskTp: property.MppTaskType, ExpectedCnt: expCnt, MPPPartitionTp: property.AnyType, CTEProducerStatus: prop.CTEProducerStatus}
+		}
+	} else {
+		lPartitionKeys, rPartitionKeys := p.GetPotentialPartitionKeys()
+		if prop.MPPPartitionTp == property.HashType {
+			var matches []int
+			switch p.JoinType {
+			case base.InnerJoin:
+				if matches = prop.IsSubsetOf(lPartitionKeys); len(matches) == 0 {
+					matches = prop.IsSubsetOf(rPartitionKeys)
+				}
+			case base.RightOuterJoin:
+				// for right out join, only the right partition keys can possibly matches the prop, because
+				// the left partition keys will generate NULL values randomly
+				// todo maybe we can add a null-sensitive flag in the MPPPartitionColumn to indicate whether the partition column is
+				//  null-sensitive(used in aggregation) or null-insensitive(used in join)
+				matches = prop.IsSubsetOf(rPartitionKeys)
+			default:
+				// for left out join, only the left partition keys can possibly matches the prop, because
+				// the right partition keys will generate NULL values randomly
+				// for semi/anti semi/left out semi/anti left out semi join, only left partition keys are returned,
+				// so just check the left partition keys
+				matches = prop.IsSubsetOf(lPartitionKeys)
+			}
+			if len(matches) == 0 {
+				return nil
+			}
+			lPartitionKeys = property.ChoosePartitionKeys(lPartitionKeys, matches)
+			rPartitionKeys = property.ChoosePartitionKeys(rPartitionKeys, matches)
+		}
+		childrenProps[0] = &property.PhysicalProperty{TaskTp: property.MppTaskType, ExpectedCnt: math.MaxFloat64, MPPPartitionTp: property.HashType, MPPPartitionCols: lPartitionKeys, CanAddEnforcer: true, CTEProducerStatus: prop.CTEProducerStatus}
+		childrenProps[1] = &property.PhysicalProperty{TaskTp: property.MppTaskType, ExpectedCnt: math.MaxFloat64, MPPPartitionTp: property.HashType, MPPPartitionCols: rPartitionKeys, CanAddEnforcer: true, CTEProducerStatus: prop.CTEProducerStatus}
+	}
+	join := PhysicalHashJoin{
+		BasePhysicalJoin:  baseJoin,
+		Concurrency:       uint(p.SCtx().GetSessionVars().CopTiFlashConcurrencyFactor),
+		EqualConditions:   p.EqualConditions,
+		NAEqualConditions: p.NAEQConditions,
+		StoreTp:           kv.TiFlash,
+		MppShuffleJoin:    !useBCJ,
+		// Mpp Join has quite heavy cost. Even limit might not suspend it in time, so we don't scale the count.
+	}.Init(p.SCtx(), p.StatsInfo(), p.QueryBlockOffset(), childrenProps...)
+	join.SetSchema(p.Schema())
+	return []base.PhysicalPlan{join}
+}
+
+func getHashJoins(super base.LogicalPlan, prop *property.PhysicalProperty) (joins []base.PhysicalPlan, forced bool) {
+	ge, p := base.GetGEAndLogical[*logicalop.LogicalJoin](super)
+	if !prop.IsSortItemEmpty() { // hash join doesn't promise any orders
+		return
+	}
+
+	forceLeftToBuild := ((p.PreferJoinType & h.PreferLeftAsHJBuild) > 0) || ((p.PreferJoinType & h.PreferRightAsHJProbe) > 0)
+	forceRightToBuild := ((p.PreferJoinType & h.PreferRightAsHJBuild) > 0) || ((p.PreferJoinType & h.PreferLeftAsHJProbe) > 0)
+	if forceLeftToBuild && forceRightToBuild {
+		p.SCtx().GetSessionVars().StmtCtx.SetHintWarning("Conflicting HASH_JOIN_BUILD and HASH_JOIN_PROBE hints detected. " +
+			"Both sides cannot be specified to use the same table. Please review the hints")
+		forceLeftToBuild = false
+		forceRightToBuild = false
+	}
+	joins = make([]base.PhysicalPlan, 0, 2)
+	switch p.JoinType {
+	case base.SemiJoin, base.AntiSemiJoin:
+		leftJoinKeys, _, isNullEQ, _ := p.GetJoinKeys()
+		leftNAJoinKeys, _ := p.GetNAJoinKeys()
+		if p.SCtx().GetSessionVars().UseHashJoinV2 && joinversion.IsHashJoinV2Supported() && CanUseHashJoinV2(p.JoinType, leftJoinKeys, isNullEQ, leftNAJoinKeys) {
+			if !forceLeftToBuild {
+				joins = append(joins, getHashJoin(ge, p, prop, 1, false))
+			}
+			if !forceRightToBuild {
+				joins = append(joins, getHashJoin(ge, p, prop, 1, true))
+			}
+		} else {
+			joins = append(joins, getHashJoin(ge, p, prop, 1, false))
+			if forceLeftToBuild || forceRightToBuild {
+				p.SCtx().GetSessionVars().StmtCtx.SetHintWarning(fmt.Sprintf(
+					"The HASH_JOIN_BUILD and HASH_JOIN_PROBE hints are not supported for %s with hash join version 1. "+
+						"Please remove these hints",
+					p.JoinType))
+				forceLeftToBuild = false
+				forceRightToBuild = false
+			}
+		}
+	case base.LeftOuterSemiJoin, base.AntiLeftOuterSemiJoin:
+		joins = append(joins, getHashJoin(ge, p, prop, 1, false))
+		if forceLeftToBuild || forceRightToBuild {
+			p.SCtx().GetSessionVars().StmtCtx.SetHintWarning(fmt.Sprintf(
+				"HASH_JOIN_BUILD and HASH_JOIN_PROBE hints are not supported for %s because the build side is fixed. "+
+					"Please remove these hints",
+				p.JoinType))
+			forceLeftToBuild = false
+			forceRightToBuild = false
+		}
+	case base.LeftOuterJoin:
+		if !forceLeftToBuild {
+			joins = append(joins, getHashJoin(ge, p, prop, 1, false))
+		}
+		if !forceRightToBuild {
+			joins = append(joins, getHashJoin(ge, p, prop, 1, true))
+		}
+	case base.RightOuterJoin:
+		if !forceLeftToBuild {
+			joins = append(joins, getHashJoin(ge, p, prop, 0, true))
+		}
+		if !forceRightToBuild {
+			joins = append(joins, getHashJoin(ge, p, prop, 0, false))
+		}
+	case base.InnerJoin:
+		if forceLeftToBuild {
+			joins = append(joins, getHashJoin(ge, p, prop, 0, false))
+		} else if forceRightToBuild {
+			joins = append(joins, getHashJoin(ge, p, prop, 1, false))
+		} else {
+			joins = append(joins, getHashJoin(ge, p, prop, 1, false))
+			joins = append(joins, getHashJoin(ge, p, prop, 0, false))
+		}
+	}
+
+	forced = (p.PreferJoinType&h.PreferHashJoin > 0) || forceLeftToBuild || forceRightToBuild
+	shouldSkipHashJoin := ShouldSkipHashJoin(p)
+	if !forced && shouldSkipHashJoin {
+		return nil, false
+	} else if forced && shouldSkipHashJoin {
+		p.SCtx().GetSessionVars().StmtCtx.SetHintWarning(
+			"A conflict between the HASH_JOIN hint and the NO_HASH_JOIN hint, " +
+				"or the tidb_opt_enable_hash_join system variable, the HASH_JOIN hint will take precedence.")
+	}
+	return
+}
+
+func getHashJoin(ge base.GroupExpression, p *logicalop.LogicalJoin, prop *property.PhysicalProperty, innerIdx int, useOuterToBuild bool) *PhysicalHashJoin {
+	var stats0, stats1 *property.StatsInfo
+	if ge != nil {
+		stats0, stats1, _, _ = ge.GetJoinChildStatsAndSchema()
+	} else {
+		stats0, stats1, _, _ = p.GetJoinChildStatsAndSchema()
+	}
+	chReqProps := make([]*property.PhysicalProperty, 2)
+	chReqProps[innerIdx] = &property.PhysicalProperty{ExpectedCnt: math.MaxFloat64, CTEProducerStatus: prop.CTEProducerStatus, NoCopPushDown: prop.NoCopPushDown}
+	chReqProps[1-innerIdx] = &property.PhysicalProperty{ExpectedCnt: math.MaxFloat64, CTEProducerStatus: prop.CTEProducerStatus, NoCopPushDown: prop.NoCopPushDown}
+	var outerStats *property.StatsInfo
+	if 1-innerIdx == 0 {
+		outerStats = stats0
+	} else {
+		outerStats = stats1
+	}
+	if prop.ExpectedCnt < p.StatsInfo().RowCount {
+		expCntScale := prop.ExpectedCnt / p.StatsInfo().RowCount
+		chReqProps[1-innerIdx].ExpectedCnt = outerStats.RowCount * expCntScale
+	}
+	hashJoin := NewPhysicalHashJoin(p, innerIdx, useOuterToBuild, p.StatsInfo().ScaleByExpectCnt(p.SCtx().GetSessionVars(), prop.ExpectedCnt), chReqProps...)
+	hashJoin.SetSchema(p.Schema())
+	return hashJoin
+}
+
+// GetHashJoin is public for cascades planner.
+func GetHashJoin(ge base.GroupExpression, la *logicalop.LogicalApply, prop *property.PhysicalProperty) *PhysicalHashJoin {
+	return getHashJoin(ge, &la.LogicalJoin, prop, 1, false)
 }
